Refactor search text state management
Removed redundant state and debounced handling from the `App` component and introduced a new `SearchTextContextProvider` to manage search-related state. This encapsulation allows for a cleaner separation of concerns and a single source of truth for the search text and its debounced value across components. It also simplifies the `SearchForm` component by removing the need to pass down search state through props. The addition of a custom hook `useSearchTextContext` ensures that the search state can be easily accessed within any component that requires it. This change aims to improve the app's maintainability and to prepare for potential future features that require search text state.